최종 수정본입니다

+++ routes/index.js
@@ -8,12 +8,15 @@ const customModals = require("../modal")
 
 router.get('/', async (req, res, next) => {
 	// 유저 목록 검색 (1)
-	const users = await libKakaoWork.getUserList();
+	console.log("inside get");
+	const users = await libKakaoWork.getAllUserList();	
+	console.log(users);
+	
 	// 검색된 모든 유저에게 각각 채팅방 생성 (2)
 	const conversations = await Promise.all(
 		users.map((user) => libKakaoWork.openConversations({ userId: user.id }))
 	);
-
+	console.log("conversations:: ", conversations);
 	// 생성된 채팅방에 메세지 전송 (3)
 	const messages = await Promise.all([
 		conversations.map((conversation) =>
@@ -24,7 +27,9 @@ router.get('/', async (req, res, next) => {
 			})
 		),
 	]);
-
+	
+	console.log("messages:: ", messages);
+	
 	res.json({
 		users,
 		conversations,
@@ -61,9 +66,9 @@ router.post('/request', async (req, res, next) => {
 router.post('/callback', async (req, res, next) => {
 	const { message, action_name, actions, action_time, value, react_user_id } = req.body; // 설문조사 결과 확인 (2)
 	const newuser = await libKakaoWork.getUserInfo(react_user_id);
-    
+    console.log("callback value::", value);
 	if(value) {	 
-		if(value === "reboot"){
+		if(value == "reboot"){
 			libKakaoWork.sendMessage({
 				conversationId: message.conversation_id,
 				text: '나의 거북씨 다시 만나기',
@@ -84,28 +89,30 @@ router.post('/callback', async (req, res, next) => {
 	}
 	
 	// after dynamic block implementation
-	// 1, 2, 3 : levels | 0: service disable
-	switch(actions.select_turtle){
-		case '1':
-		case '2':
-		case '3':
-			newuser.level = actions.select_turtle;
-			newuser.NextUpdateTime = Date.now();
-			newuser.startTime = actions.start_time *= 1;
-			newuser.endTime = actions.end_time *= 1;
-			newuser.working = 0;
-			userService.updateNextTime(newuser);
-			userService.checkAndAppendUser(newuser, message);	
-			break;
-		case '0':
-			userService.deleteUser(newuser);	  
-			await libKakaoWork.sendMessage({
-				conversationId: message.conversation_id,
-				text: '안녕하세요, 친절한 거북씨에요',
-				blocks: customModals.registerRejectModal.blocks
-			});  
-			break;
-		default:
+	// 1, 2, 3 : levels | 0: service disabl
+	if(actions){
+		switch(actions.select_turtle){
+			case '1':
+			case '2':
+			case '3':
+				newuser.level = actions.select_turtle;
+				newuser.NextUpdateTime = Date.now();
+				newuser.startTime = actions.start_time *= 1;
+				newuser.endTime = actions.end_time *= 1;
+				newuser.working = 0;
+				userService.updateNextTime(newuser);
+				userService.checkAndAppendUser(newuser, message);	
+				break;
+			case '0':
+				userService.deleteUser(newuser);	  
+				await libKakaoWork.sendMessage({
+					conversationId: message.conversation_id,
+					text: '안녕하세요, 친절한 거북씨에요',
+					blocks: customModals.registerRejectModal.blocks
+				});  
+				break;
+			default:
+		}
 	}
 	
 	res.json({ result: true });
